% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_lookup.R
\name{team_lookup}
\alias{team_lookup}
\title{Lookup table for NCAA teams}
\usage{
team_lookup(team, team_db = D1VB_team_names)
}
\arguments{
\item{team}{a string containing the name of the team to look up}

\item{team_db}{a data frame containing the names of teams (and conferences) used by different organizations. For columns containing team names, the organization name should be the column name. 
The default data frame is the D1VB_team_names dataset containing NCAA, Volleymetrics, and Rich Kern team names for 335 D1 women's volleyball teams.}
}
\value{
A subset of `team_db` containing possible matches for the team of interest
}
\description{
Finds likely match(es) for desired volleyball team or conference.
    The NCAA, Volleymetrics, and Rich Kern all use a different naming scheme for teams. This function, combined with an appropriate data table,
allows the user to find the equivalent names used by all three organizations given a guess at the name used by any organization.
    In addition, the default data table used allows the same function to be used to look up all teams in a given conference.
}
