% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_add_cover.R
\name{vwp_add_cover}
\alias{vwp_add_cover}
\title{Add cover}
\usage{
vwp_add_cover(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same data frame, with an "add_cover" variable indicating whether the touch was a cover.
Also, replaces the start zone of the touch if the touch was a cover.
}
\description{
Adds whether a dig was a cover. #'
}
