% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_attack_markov_chain.R
\name{vwp_attack_markov_chain}
\alias{vwp_attack_markov_chain}
\title{Attack Markov Chain}
\usage{
vwp_attack_markov_chain(plays, system = TRUE)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}

\item{system}{whether to break down by in vs. out-of-system}
}
\value{
A data frame indicating the modeled point-win-probability of an in-system, out-of-system, or no-set attack
}
\description{
Estimates transition matrix between attack types
}
