% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_attack_start_zone.R
\name{vwp_attack_start_zone}
\alias{vwp_attack_start_zone}
\title{Assign start zone of attack}
\usage{
vwp_attack_start_zone(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same data frame, with an "attack_start_zone" variable indicating the starting zone of the (preceding) attack
}
\description{
Assigns a code to attacks and digs to indicate the start zone of the attack. Currently only supports the start_zone column (no subzones yet).
}
