% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_attack_type.R
\name{vwp_attack_type}
\alias{vwp_attack_type}
\title{Assign attack type}
\usage{
vwp_attack_type(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same data frame, with an "attack_type" variable indicating whether the attack was a hard spike or offspeed attack (tip, roll shot, etc.).
}
\description{
Assigns a code to attacks, blocks and digs to indicate the quality of the attack. Currently just recodes hard spikes vs. everything else is off-speed.
}
