% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_block_front_row.R
\name{vwp_block_front_row}
\alias{vwp_block_front_row}
\title{Add opposing players in front row during an attack}
\usage{
vwp_block_front_row(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same data frame, with three additional columns ("block_front_left", "block_front_center", "block_front_right") indicating the opposing team's front line during an attack
}
\description{
Using rotation information, record the names of the players who start in the front left/center/right positions, for the blocking team.
Because teams may run unusual blocking schemes (switch-blocking, middle on a pin, etc.), we can only know which players could possibly have been in position to block an attack.
Furthermore, a player may influence the attack without directly being involved in the block (e.g., if the setter intentionally sets away from a pin blocker)
}
