% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_block_touched.R
\name{vwp_block_touched}
\alias{vwp_block_touched}
\title{Add whether the dig is off a block touch or not}
\usage{
vwp_block_touched(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same data frame, with a "block_touched" variable taking the value 1 if the touch is a dig/cover off a block touch and 0 otherwise
}
\description{
Classify a dig as being off a block touch or directly off an attack. Covers should always be classified as off a block touch.
}
