% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_blockers.R
\name{vwp_blockers}
\alias{vwp_blockers}
\title{Assign quality of block}
\usage{
vwp_blockers(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same data frame, with a "blockers" variable indicating whether the attack faced a solo, double, triple, or seam block, or no block
}
\description{
Assigns a code to attacks and blocks to indicate the quality of the block. Currently just recodes the num_players column.
}
