% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_dig_distance.R
\name{vwp_dig_distance}
\alias{vwp_dig_distance}
\title{Dig distance}
\usage{
vwp_dig_distance(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data.}
}
\value{
The same data frame, with the estimated (court) distance traveled by the ball between the attack and the dig
}
\description{
Estimates the distance traveled by an attack to a dig attempt
}
\details{
We do not take into account deflections from a block when computing distance, it's just a
straight-line distance between the attack location and the dig location
}
