% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_extract_relevant_features.R
\name{vwp_extract_relevant_features}
\alias{vwp_extract_relevant_features}
\title{Extract relevant features for a Volleyball Win Probability Model}
\usage{
vwp_extract_relevant_features(
  plays,
 
    features = c("team, opponent, match_id, point_id, video_time, player_name,\\n                                                             skill, skill_type, evaluation_code, evaluation, attack_code, set_code,\\n                                                             start_zone, end_zone, end_subzone, skill_subtype, num_players,\\n                                                             home_p2, home_p3, home_p4, visiting_p2, visiting_p3, visiting_p4,\\n                                                             home_team, visiting_team, point_won_by, serving_team, phase")
)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data.}

\item{features}{a character vector containing the names of the variables to be included. If you don't know which ones you want to keep in, use the default.
This should work even if your plays object does not contain some of these columns. If you include too few features, you may accidentally delete rows corresponding to different touches.
Also, you may not be able to run the full model later as this function is designed for pre-processing.}
}
\value{
The same data frame, with duplicate rows and extraneous columns removed
}
\description{
Subsets a data frame to include only features of interest and remove any duplicate rows (usually games that got added twice)
}
