% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_fix_skill_codes.R
\name{vwp_fix_skill_codes}
\alias{vwp_fix_skill_codes}
\title{Fix clearly impossible skill codes}
\usage{
vwp_fix_skill_codes(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same data frame, with the "skill" variable updated to change clearly-impossible touches to their more likely equivalents
}
\description{
Recodes touches that can't possibly have been classified correctly in the original datavolley file.
}
