% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_compute_pwp_workflow.R
\name{vwp_input_pwp}
\alias{vwp_input_pwp}
\title{Compute input point-win-probability}
\usage{
vwp_input_pwp(design_matrix, coef_matrix, PWP_matrix)
}
\arguments{
\item{design_matrix}{a n x p matrix where n is the number of touches to model and p is the number of features, usually computed from `vwp_model_matrix()`.}

\item{coef_matrix}{a p x 8 matrix containing the coefficients of a multinomial logistic regression model estimating the odds of each of 8 possible next attack outcomes
(OI, ON, OO, OW, TI, TN, TO, TW in that order) where the first letter represents team or opponent result and the second letter represents
the type of attack (I = in-system, N = net play, O = out-of-system, W = win/lose point without another attack being made).}

\item{PWP_matrix}{an 8 x 1 matrix containing the modeled point-win probability values for each of the 8 outcomes in the same order
as they are listed in the coef_matrix. The idea is to take a weighted average of the point-win probability values for each type of next attack,
with the weights being the probability of each outcome.}
}
\value{
A numeric vector estimating the overall point-win probability of a set of touches.
}
\description{
The input point-win probability for a touch is the modeled probability of winning the point at the exact instant before the ball is contacted.
Input point-win probability is estimated from model parameters that are available just prior to the touch (usually location, and information about the previous touch).
}
