% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_kcode.R
\name{vwp_kcode}
\alias{vwp_kcode}
\title{K-Code Lookup}
\usage{
vwp_kcode(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same data frame, with a "system" variable indicating an in-system, out-of-system, or no set
}
\description{
Determines if the attack included a middle on a route (K combination) or not
}
