% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_multinomial_regression.R
\name{vwp_multinomial_regression}
\alias{vwp_multinomial_regression}
\title{Volleyball Win Probability Multinomial Regression Model}
\usage{
vwp_multinomial_regression(plays, formula, nmax = 2e+05, seed = 100)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data.}

\item{formula}{a formula whose left-hand side contains the name of the variable in `plays` containing the next attack, and 
whose right-hand side contains the features used in the regression}

\item{nmax}{the maximum number of rows to use when fitting the model. Very large datasets can cause memory issues. If you are 
getting error messages related to an inability to allocate memory, set nmax to a smaller number. By default we use 200000.}

\item{seed}{a seed for reproducibility of randomly sampling very large datasets.}
}
\value{
A coefficient matrix for a multinomial regression model, where each column represents one possible next attack and 
each row represents the contribution to the odds ratio of that outcome vs. the team winning the point without any more attacks made. The matrix has dimensions
p x 8, where p is the number of features (including dummary variables) in the model matrix.
}
\description{
Builds a model to predict the next attack from a touch and outputs the coefficient matrix
}
