% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_next_attack_details.R
\name{vwp_next_attack_details}
\alias{vwp_next_attack_details}
\title{Assign next attack}
\usage{
vwp_next_attack_details(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same data frame, with a "blockers" variable indicating whether the attack faced a solo, double, triple, or seam block, or no block
}
\description{
Assigns a code to all skills indicating the team and system of the next attack
}
