% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_player_number_to_player_name.R
\name{vwp_player_number_to_player_name}
\alias{vwp_player_number_to_player_name}
\title{Convert player number to player name}
\usage{
vwp_player_number_to_player_name(player_number, player_team, plays)
}
\arguments{
\item{player_number}{the number of the player to replace}

\item{player_team}{the team for which the player plays}

\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same data frame, with an "attack_start_zone" variable indicating the starting zone of the (preceding) attack
}
\description{
Looks up a player's team and number in the play-by-play data and reports the corresponding name
}
