% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_receiving_player.R
\name{vwp_receiving_player}
\alias{vwp_receiving_player}
\title{Assign receiving player}
\usage{
vwp_receiving_player(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same data frame, with a "receiving_player" variable indicating which player (if any) received the serve
}
\description{
Assigns the receiving player to the serve. If the serve is not received (e.g., error), assigns TEAM as the receiving player.
   The model does not compute a separate point-win probability for a reception, so this can be used to assign credit for receptions.
   Additionally, adds a receiving player TEAM for unforced errors to make the win-probability differences balance out.
}
