% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_set_xy.R
\name{vwp_set_xy}
\alias{vwp_set_xy}
\title{Transform volleyball Set x-y coordinates}
\usage{
vwp_set_xy(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data.}
}
\value{
The same object, with new columns containing the transformed x and y coordinates of the set
}
\description{
Transforms set x-y coordinates based on the distance from Zone 3
 
We assume that there are negligible effects of passing slightly left/right/off the net compared to perfect pass. 
However, if the pass is outside Zone 3, the setting options and quality decrease quickly the further outside the zone. 
We transform x and y coordinates to their distance from Zone 3.
}
