% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_set_zone.R
\name{vwp_set_zone}
\alias{vwp_set_zone}
\title{Assign quality of set location.}
\usage{
vwp_set_zone(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same data frame, with a "set_zone" variable indicating the location of the (previous) set.
}
\description{
Assigns a code to attacks and sets to indicate the quality of the set location. Currently just recodes the end zone of the (previous) set.
}
