% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_setter_touch.R
\name{vwp_setter_touch}
\alias{vwp_setter_touch}
\title{Add a setter flag}
\usage{
vwp_setter_touch(plays, setters)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data.}

\item{setters}{a character vector containing the names of all setters on the teams recorded in the `plays` dataset.
For smaller datasets, the user will typically know this. For larger datasets, it must be inferred.}
}
\value{
The same dataset, with a column with value 1 if the player touching the ball is a setter and 0 otherwise
}
\description{
Indicates whether the setter touched the ball
}
\details{
Due to the nature of the positions, attacks, digs/freeballs, and sets from setters are qualitatively different from these touches
when played by other positions. For teams running a 6-2 offense, this function may still be worthwhile for digs and sets.
}
