% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_time_to_touch.R
\name{vwp_time_to_touch}
\alias{vwp_time_to_touch}
\title{Assign time from previous touch}
\usage{
vwp_time_to_touch(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same data frame, with a "time_to_touch" variable indicating how many seconds (on video time) since the previous touch
}
\description{
Computes the time from the previous touch until the next touch.
}
