% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwp_zone_distance.R
\name{vwp_zone_distance}
\alias{vwp_zone_distance}
\title{Compute zone distance}
\usage{
vwp_zone_distance(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same data frame, with a "distance" variable indicating the estimated court distance the ball traveled between the attack and the dig
}
\description{
Assigns a numerical value to the estimated distance between the starting and ending location of an attack that is dug.
Values are in meters. Height of the attack is not considered.
    Currently assumes that all attacks start at the very front center of the zone and are dug in the exact center of the zone.
}
